chore: Refactor animated radio implementation in device found flow
